Refactor wallet and cryptographic operations to use mona-crypto library
- Updated tests to import key functions from key_old module.
- Modified Cargo.toml files to include mona-crypto as a workspace dependency.
- Refactored main.rs and accounts.rs to utilize mona-crypto for wallet operations.
- Enhanced stake module to load wallets using mona-crypto.
- Updated blockchain module to verify transaction signatures using mona-crypto.
- Added security features and best practices to the mona-crypto README.
- Improved encryption and decryption functions with input validation.
- Implemented secure memory handling and constant-time comparisons for signature verification.
- Streamlined transaction signing and verification processes in panorama module.
- Added comprehensive error handling and logging throughout the codebase.
