+++ Cargo.toml
@@ -93,7 +93,7 @@ crypto = "0.5.1"
 digest = "0.10.7"
 hex = "0.4.3"
 rand = "0.8.5"  # version 0.8.5 is not compatible with the move-package crate
-sha2 = "0.10.8"
+sha2 = "0.10.9"
 sha3 = "0.10.8"
 uuid = { version = "1.15.0", features = ["v4", "serde"] }
 
@@ -105,8 +105,8 @@ serde = { version = "1.0.217", features = ["derive", "rc"] }
 serde_json = { version = "1.0.134", features = ["preserve_order", "arbitrary_precision"] }
 serde_yaml = "0.9.34-deprecated"
 semver = { version = "1.0.26" }
-toml = "0.8.20"
-toml_edit = "0.22.24"
+toml = "0.8.22"
+toml_edit = "0.22.26"
 
 # Async & Network Programming
 async-trait = "0.1"
