+++ crates/command/Cargo.toml
@@ -22,6 +22,9 @@ panorama.workspace = true
 kari-move.workspace = true
 mona-blockchain.workspace = true
 mona-storage.workspace = true
+mona-crypto.workspace = true
+
+
 tokio.workspace = true
 serde_json.workspace = true
 serde.workspace = true
