+++ crates/rpc/rpc-api/Cargo.toml
@@ -33,5 +33,6 @@ mona-storage = { workspace = true }
 mona-blockchain = { workspace = true }
 move-core-types = { workspace = true }
 mona-vm = { workspace = true }
+mona-crypto = { workspace = true }
 key = { workspace = true }
 tokio = { workspace = true, features = ["full"] }
