+++ crates/rpc/rpc-api/src/accounts.rs
@@ -1,6 +1,7 @@
 use std::str::FromStr;
 use jsonrpc_core::{Params, Result as JsonRpcResult, Error as RpcError, ErrorCode};
 
+use mona_crypto::{list_wallet_files, load_wallet};
 use mona_types::address::Address;
 use mona_blockchain::{blockchain::{load_blockchain_with_retry, BALANCES, BLOCKCHAIN_DATA}, chain_id::CHAIN_ID};
 use panorama::simulation::process_transfer;
@@ -12,7 +13,7 @@ use serde_json::{json, Value as JsonValue};
 use serde::Deserialize;
 
 use tokio::sync::mpsc;
-use key::{load_wallet, list_wallet_files};
+
