+++ monaos/mona-blockchain/src/block.rs
@@ -2,6 +2,8 @@ use crate::chain_id::CHAIN_ID;
 use consensus_pos::HashAlgorithm;
 use serde::{Deserialize, Serialize};
 use mona_types::address::Address;
+use mona_crypto::verify_signature; // Add mona-crypto dependency
+use log;
 
 // Define the Transaction struct
 #[derive(Serialize, Deserialize, Clone)]
@@ -36,57 +38,28 @@ impl Transaction {
     }
 
     // Verify the transaction signature
-    pub fn verify_signature(&self) -> Result<bool, Box<dyn std::error::Error>> {
+    pub fn verify(&self) -> bool {
+        // Check if signature exists
         if self.signature.is_empty() {
-            log::debug!("Transaction has empty signature, cannot verify");
-            return Ok(false); // Can't verify an empty signature
+            log::warn!("Transaction {} has no signature", self.transaction_id);
+            return false;
         }
-        
+
+        // Generate the message that was originally signed
         let message = self.to_signable_message();
-        let address = self.sender.to_hex_literal();
         
-        log::debug!("Verifying signature for: address={}, message_len={}, sig_len={}", 
-                   address, message.len(), self.signature.len());
-        
-        // Try general verification first
-        match key::verify_signature(&address, &message, &self.signature) {
-            Ok(true) => {
-                log::debug!("Signature verification succeeded with general method");
-                return Ok(true);
+        // Verify signature using mona-crypto
+        match verify_signature(&self.sender.to_string(), &message, &self.signature) {
+            Ok(is_valid) => {
+                if !is_valid {
+                    log::warn!("Invalid signature for transaction {}", self.transaction_id);
+                }
+                is_valid
             },
-            _ => {
-                log::debug!("General verification failed, trying specific curve types");
-            }
-        }
-        
-        // Try verification with K256 first
-        match key::verify_signature_k256(address.trim_start_matches("0x"), &message, &self.signature) {
-            Ok(true) => {
-                log::debug!("K256 signature verification succeeded");
-                return Ok(true);
-            },
-            Err(e) => {
-                log::debug!("K256 verification error: {}", e);
-            },
-            _ => {
-                log::debug!("K256 verification returned false");
-            }
-        }
-        
-        // If K256 verification fails, try P256
-        match key::verify_signature_p256(address.trim_start_matches("0x"), &message, &self.signature) {
-            Ok(true) => {
-                log::debug!("P256 signature verification succeeded");
-                Ok(true)
-            },
-            Ok(false) => {
-                log::debug!("P256 verification returned false");
-                Ok(false)
-            },
-            Err(e) => {
-                log::warn!("P256 verification failed: {}", e);
-                // We'll return false rather than an error to keep processing flowing
-                Ok(false)
+            Err(err) => {
+                log::error!("Error verifying signature for transaction {}: {}", 
+                    self.transaction_id, err);
+                false
             }
         }
     }
