+++ monaos/mona-crypto/Cargo.toml
@@ -24,19 +24,20 @@ argon2 = "0.5"
 bip39 = { version = "2.0", features = ["rand"] }
 
 # Serialization
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-serde_yaml = "0.9"
-bincode = "1.3.3"
+serde = { version = "1.0.219", features = ["derive"] }
+serde_json = "1.0.140"
+serde_yaml = "0.9.34-deprecated"
+bincode = "2.0.1"
 toml = "0.8.22"
 hex = "0.4.3"
 
 # Utilities
 thiserror = "2.0.12"
-rand = "0.8.5"
+rand = "0.9.1"
 log = "0.4.27"
 base64 = "0.22.1"
 
 # Project dependencies
 mona-types.workspace = true
 common.workspace =true
+key.workspace = true
