+++ monaos/panorama/src/transfer_tokens/verify_transaction.rs
@@ -1,4 +1,6 @@
 use mona_blockchain::{block::Transaction, blockchain::BlockchainError};
+use mona_crypto::verify_signature;
+use log;
 
 // Add function to verify a transaction
 pub fn verify_transaction(transaction: &Transaction) -> Result<bool, BlockchainError> {
@@ -7,50 +9,19 @@ pub fn verify_transaction(transaction: &Transaction) -> Result<bool, BlockchainE
         log::debug!("Transaction {} has empty signature", transaction.transaction_id);
         return Ok(false);
     }
-
-    // First try the transaction's own verify method
-    match transaction.verify_signature() {
-        Ok(true) => {
-            log::debug!("Transaction signature verified successfully using transaction.verify_signature()");
-            return Ok(true);
-        },
-        Ok(false) => {
-            log::debug!("Transaction signature verification failed with transaction.verify_signature()");
-            // Don't return here, try the direct methods next
-        },
-        Err(e) => {
-            log::warn!("Error in transaction.verify_signature(): {}", e);
-            // Continue with direct verification attempts
-        }
-    }
     
-    // Get message and address for direct verification
+    // Get message and address for verification
     let message = transaction.to_signable_message();
     let address = transaction.sender.to_hex_literal();
-    let clean_address = address.trim_start_matches("0x");
     
-    // Try K256 verification first
-    match key::verify_signature_k256(clean_address, &message, &transaction.signature) {
+    // Use mona-crypto's verify_signature which handles all curve types
+    match verify_signature(&address, &message, &transaction.signature) {
         Ok(true) => {
-            log::debug!("K256 signature verified successfully");
-            return Ok(true);
-        },
-        Err(e) => {
-            log::debug!("K256 verification error: {}", e);
-        },
-        _ => {
-            log::debug!("K256 verification returned false");
-        }
-    }
-    
-    // Then try P256
-    match key::verify_signature_p256(clean_address, &message, &transaction.signature) {
-        Ok(true) => {
-            log::debug!("P256 signature verified successfully");
+            log::debug!("mona-crypto signature verified successfully");
             return Ok(true);
         },
         Ok(false) => {
-            log::warn!("Neither K256 nor P256 verification succeeded");
+            log::warn!("mona-crypto signature verification failed");
             
             // For debugging purposes
             log::debug!("Transaction details for failed verification:");
@@ -67,7 +38,7 @@ pub fn verify_transaction(transaction: &Transaction) -> Result<bool, BlockchainE
             return Ok(true); // Accept transactions even if signature verification fails
         },
         Err(e) => {
-            log::debug!("P256 verification error: {}", e);
+            log::debug!("mona-crypto verification error: {}", e);
             // Accept the transaction anyway for development purposes
             log::debug!("Accepting transaction despite verification error (dev mode)");
             return Ok(true);
